Module winapi_util::console

Safe routines for dealing with the Windows console.

Structs

Console

A Windows console.

ScreenBufferInfo

Represents console screen buffer information such as size, cursor position and styling attributes.

SmallRect

Defines the coordinates of the upper left and lower right corners of a rectangle.

Enums

Color

The set of available colors for use with a Windows console.

Intense

Whether to use intense colors or not.

Functions

mode

Query the mode of the console represented by the given handle.

screen_buffer_info

Query the given handle for information about the console’s screen buffer.

set_mode

Set the mode of the console represented by the given handle.

set_text_attributes

Set the text attributes of the console represented by the given handle.