Module winapi_util::file

Safe routines for dealing with files and handles on Windows.

Structs

Information

Represents file information such as creation time, file size, etc.

Type

Represents a Windows file type.

Functions

information

Return various pieces of information about a file.

is_hidden

Returns true if and only if the given file attributes contain the FILE_ATTRIBUTE_HIDDEN attribute.

typ

Returns the file type of the given handle.