Struct winapi_util::file::Information

pub struct Information(_);

Represents file information such as creation time, file size, etc.

This wraps a BY_HANDLE_FILE_INFORMATION.

Implementations

impl Information

pub fn file_attributes(&self) -> u64

Returns file attributes.

This corresponds to dwFileAttributes.

pub fn is_hidden(&self) -> bool

Returns true if and only if this file information has the FILE_ATTRIBUTE_HIDDEN attribute.

pub fn creation_time(&self) -> Option<u64>

Return the creation time, if one exists.

This corresponds to ftCreationTime.

pub fn last_access_time(&self) -> Option<u64>

Return the last access time, if one exists.

This corresponds to ftLastAccessTime.

pub fn last_write_time(&self) -> Option<u64>

Return the last write time, if one exists.

This corresponds to ftLastWriteTime.

pub fn volume_serial_number(&self) -> u64

Return the serial number of the volume that the file is on.

This corresponds to dwVolumeSerialNumber.

pub fn file_size(&self) -> u64

Return the file size, in bytes.

This corresponds to nFileSizeHigh and nFileSizeLow.

pub fn number_of_links(&self) -> u64

Return the number of links to this file.

This corresponds to nNumberOfLinks.

pub fn file_index(&self) -> u64

Return the index of this file. The index of a file is a purpotedly unique identifier for a file within a particular volume.