winapi_util

Struct Handle

pub struct Handle(/* private fields */);

A handle represents an owned and valid Windows handle to a file-like object.

When an owned handle is dropped, then the underlying raw handle is closed. To get a borrowed handle, use HandleRef.

Implementations

impl Handle

pub fn from_file(file: File) -> Handle

Create an owned handle to the given file.

When the returned handle is dropped, the file is closed.

Note that if the given file represents a handle to a directory, then it is generally required that it have been opened with the FILE_FLAG_BACKUP_SEMANTICS flag in order to use it in various calls such as information or typ. To have this done automatically for you, use the from_path_any constructor.

pub fn from_path<P: AsRef<Path>>(path: P) -> Result<Handle>

Open a file to the given file path, and return an owned handle to that file.

When the returned handle is dropped, the file is closed.

If there was a problem opening the file, then the corresponding error is returned.

pub fn from_path_any<P: AsRef<Path>>(path: P) -> Result<Handle>

Like from_path, but supports opening directory handles as well.

If you use from_path on a directory, then subsequent queries using that handle will fail.

pub fn as_file(&self) -> &File

Return this handle as a standard File reference.

pub fn as_file_mut(&mut self) -> &mut File

Return this handle as a standard File mutable reference.