Trait winapi_util::AsHandleRef

pub trait AsHandleRef {
    fn as_handle_ref(&self) -> HandleRef;

    fn as_raw(&self) -> RawHandle { ... }
}

Construct borrowed and valid Windows handles from file-like objects.

Required Methods

fn as_handle_ref(&self) -> HandleRef

A borrowed handle that wraps the raw handle of the Self object.

Provided Methods

fn as_raw(&self) -> RawHandle

A convenience routine for extracting a HandleRef from Self, and then extracting a raw handle from the HandleRef.